# By let4be
# Via let4be
* 'develop' of github.com:lavab/web:
  clean up
  styles setting: ask removed as it's hard to implement technically(too time consuming)
  added "styles setting" - none, all, ask. todo: implement ask
  #462, added custom title
  closed #484, added confirmations for password change, ls disable and keys removal(can serve as example for #462)
  proper i18n for dialogs :|
  added dialogs support #462
  bump 0.2.4
  thread.members LB_EMAIL_TO_YOURSELF
  fixed hidden contacts creation
  fix thread.members to respect fancy formatted emails with name and email
  closed #489, inversed contact.isSecured logic. Consider contact secured when at least one email is secured
  closed #486, added custom tagOnFocusLoss directive to asap ui-select with tagging on focus loss
  closed #485, allow sending emails with empty subject and body, replace subject with a stub if no subject found
  fixed using new key.email instead of hacking with existing key.user
  closed #481, added support for new key fields and methods
  closed #479, sort by modified date by default

+++ src/js/apps/AppLavaboom/controllers/ctrlContactList.js
@@ -1,10 +1,12 @@
-module.exports = /*@ngInject*/($rootScope, $scope, $translate, $state, $stateParams, co, contacts, user, hotkey) => {
+module.exports = /*@ngInject*/($rootScope, $scope, $translate, $state, $stateParams, dialogs, co, contacts, user, hotkey) => {
 	$scope.selectedContactId = null;
 	$scope.searchText = '';
 
 	const translations = {
 		LB_NEW_CONTACT_SHORT: '',
-		LB_EMPTY_CONTACT_SHORT: ''
+		LB_EMPTY_CONTACT_SHORT: '',
+		LB_CONFIRM: '%',
+		TITLE_CONFIRM_DELETE:''
 	};
 
 	$translate.bindAsObject(translations, 'MAIN.CONTACTS');
@@ -109,6 +111,15 @@ module.exports = /*@ngInject*/($rootScope, $scope, $translate, $state, $statePar
 	});
 
 	$scope.deleteContact = (contactId) => co(function *(){
+		const contact = contacts.getContactById(contactId);
+
+		const confirmed = yield co.def(dialogs.confirm(
+			translations.TITLE_CONFIRM_DELETE,
+			translations.LB_CONFIRM({name: contact.getFullName()})
+		).result, 'cancelled');
+		if (confirmed == 'cancelled')
+			return;
+
 		return yield contacts.deleteContact(contactId);
 	});
